+++ tests/Command/PeekTest.php
@@ -36,12 +36,16 @@ class PeekTest extends CommandTestCase
 
     public function testNotFoundThrowsException(): void
     {
+        $jobId = rand();
+
         $this->expectException(NotFoundException::class);
+        $this->expectExceptionMessage(sprintf(NotFoundException::JOB_ID_MSG_F, $jobId));
+        $this->expectExceptionCode(NotFoundException::JOB_ID_CODE);
 
         $this->socket->expects(static::any())
             ->method('read')
             ->willReturn('NOT_FOUND');
-        (new Peek(10))->process($this->socket);
+        (new Peek($jobId))->process($this->socket);
     }
 
     public function testUnknownStatusThrowsException(): void