+++ tests/Command/TouchTest.php
@@ -32,12 +32,16 @@ class TouchTest extends CommandTestCase
 
     public function testErrorThrowsException(): void
     {
+        $jobId = rand();
+
         $this->expectException(NotFoundException::class);
+        $this->expectExceptionMessage(sprintf(NotFoundException::JOB_ID_MSG_F, $jobId));
+        $this->expectExceptionCode(NotFoundException::JOB_ID_CODE);
 
         $this->socket->expects(static::any())
             ->method('read')
             ->willReturn('NOT_TOUCHED');
-        (new Touch(123))->process($this->socket);
+        (new Touch($jobId))->process($this->socket);
     }
 
     public function testUnknownStatusThrowsException(): void